requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# API Documentation
drf-yasg==1.21.7

# Environment variables
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
